refactor: move api/v1 DTOs into dedicated dto subpackage
Separates request/response records from controllers in api/v1, mirroring
the existing web/viewmodel pattern used for web-tier view models.
