+++ src/main/java/de/hdawg/rankinginfo/api/v1/dto/PlayerDetailData.java
@@ -1,4 +1,4 @@
-package de.hdawg.rankinginfo.api.v1;
+package de.hdawg.rankinginfo.api.v1.dto;
 
 import java.util.List;
