+++ src/main/java/de/hdawg/rankinginfo/api/v1/dto/PlayerDetailResponse.java
@@ -1,3 +1,3 @@
-package de.hdawg.rankinginfo.api.v1;
+package de.hdawg.rankinginfo.api.v1.dto;
 
 public record PlayerDetailResponse(PlayerDetailData data) {}
