+++ src/main/java/de/hdawg/rankinginfo/api/v1/dto/PlayerLink.java
@@ -1,3 +1,3 @@
-package de.hdawg.rankinginfo.api.v1;
+package de.hdawg.rankinginfo.api.v1.dto;
 
 public record PlayerLink(String self) {}
