+++ src/main/java/de/hdawg/rankinginfo/api/v1/dto/PlayerSearchItem.java
@@ -1,4 +1,4 @@
-package de.hdawg.rankinginfo.api.v1;
+package de.hdawg.rankinginfo.api.v1.dto;
 
 public record PlayerSearchItem(
     int dtb_id, String lastname, String firstname, String club, PlayerLink links) {}
