+++ src/main/java/de/hdawg/rankinginfo/api/v1/dto/PlayerSearchRequest.java
@@ -1,3 +1,3 @@
-package de.hdawg.rankinginfo.api.v1;
+package de.hdawg.rankinginfo.api.v1.dto;
 
 public record PlayerSearchRequest(String lastname, String yob, int total_count) {}
