+++ src/main/java/de/hdawg/rankinginfo/api/v1/dto/package-info.java
@@ -1,0 +1,2 @@
+@org.jspecify.annotations.NullMarked
+package de.hdawg.rankinginfo.api.v1.dto;
